requirements.txt
flask-sqlalchemy==2.5.1
flask-login==0.5.0
pymysql==1.0.2
pytz==2021.3
flask_wtf==1.0.0
flask_bcrypt==0.7.1
wtforms==3.0.0
datetime==4.3
flask-restful==0.3.9
flask-restful-swagger-2==0.35
flask-cors==3.0.10
google-api-python-client==2.36.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
authlib==0.15.5
email-validator==1.1.3
blinker==1.4
html5lib==1.1
requests==2.27.1
bs4==0.0.1
cython==0.29.33
pandas==1.1.5